Unify interrupt return paths from TSP into the TSPD
The TSP passes control back to EL3 through an SMC in response to an interrupt in
the following scenarios:

1. An FIQ that has preempted normal world execution is handed to the TSP by EL3
   firmware. The interrupt is deasserted or is preempted by a higher priority
   interrupt by the time the TSP acknowledges it. In this case, the TSP issues
   an SMC with the ID as TSP_EL3_FIQ.

2. An IRQ preempts TSP execution while it is handling a standard SMC. In this
   case, the TSP issues an SMC with the ID as TSP_PREEMPTED.

In both the cases, the TSPD hands control back to the normal world. In 2. it
additionally returns an error code to the normal to indicate that the standard
SMC it had issued has been preempted but not completed.

This patch unifies the handling of these two cases in the TSPD and ensures that
the TSP only uses TSP_PREEMPTED instead of separate SMC IDs. The TSPD is able to
distinguish between the two scenarios by checking if a standard SMC is in
progress.

On a GICv3 system, when the secure world has affinity routing enabled, in 2. an
FIQ will preempt TSP execution instead of an IRQ. The FIQ could be a result of a
Group 0 or a Group 1 NS interrupt. In both case, the TSPD passes control back to
the normal world upon receipt of the TSP_PREEMPTED SMC. A Group 0 interrupt will
immediately preempt execution to EL3 where it will be handled.

Change-Id: I9895344db74b188021e3f6a694701ad272fb40d4