sunxi: SCPI: add handler boilerplate
Add the basic framework to handle request via the SCPI interface. This
reads and decodes the request found in the shared memory region,
provides the function framework to handle a certain command and takes
care about passing on the return value.
This for just the basic capability call is implemented.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>