STUPID copy of plat/sun50iw1p1 to prepare hackish H6 support
Copy the plat/sun50iw1p1 directory over to plat/sun50iw6p1 to pave the
way for a quick and hackish H6 support.
This is wrong for several reasons:
- There should be a proper upstream port for the Allwinner SoCs. This is
WIP and would probably include support for the H6 from the very beginning.
- Having a new directory directly under plat/ is not the intended way of
adding support for a new SoC under an existing platform. See above.
- There are many similarities between the H6 and the A64/H5. However all
the MMIO addresses have changed, which we have hardcoded so far in the
code. To avoid some non-trivial refactoring on a dead code base, let's
just copy everything for now and hack in the changed addresses in a
follow-up patch.

Despite the hackish and wrong nature of this patch it serves the purpose
of getting something up very quickly. Also it demonstrates the
differences between the two SoCs, which helps to find the right
abstraction for a proper port.

This is a plain copy, plus adjusting the file names in platform.mk. No
actual code changes so far, so following patches show the differences
between A64/H5 and the new SoC.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>